ALS-12139: Implement partitioning, basic functionality

+++ data/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/data/phenotype/ColumnMeta.java
@@ -92,7 +92,7 @@ public class ColumnMeta implements Serializable {
         return this;
     }
 
-    public long getPatientCount() {
+    public int getPatientCount() {
         return patientCount;
     }
